fixed `TestSeeder` class not resolved
